로그인 로그 JPA 동작 완료

+++ src/main/java/com/artiplace/api/common/repository/LoginLogRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.artiplace.api.common.entity.LoginLogEntity;
 	
 @Repository
-public interface LoginLogRepository extends JpaRepository<LoginLogEntity, BigInteger> {
+public interface LoginLogRepository extends JpaRepository<LoginLogEntity, Long> {
 }
